uint -> int.

+++ shaders/terrain.frag
@@ -40,7 +40,7 @@ void main() {
 	normal_map.xyz = normalize(normal_map.xyz * 2.0 - 1.0);
 	vec4 accumLighting = originalColor * vec4(Lgt.ambient_intensity,1.f);
 
-	for(int light = 0; uint(light) < Lgt.num_lights; ++light) {
+	for(int light = 0; light < Lgt.num_lights; ++light) {
 		vec3 light_distance = Lgt.lights[light].position.xyz - position;
 		vec3 dir = normalize(light_distance);
 		vec3 light_dir;
